Detail url working in generated module.
The detail url in the backend was not working correctly; when clicked
one would land on the 404 page.  This change fixes this, with one small
caveat:
It appears the meta.base.tpl snippet is shared between add and edit
actions, this results in {$item.url} being displayed when adding an
item, however this is shortly after replaced by the ajax action
generating the slug.  This seems like an ok trade-off for the time
being, IMHO.

Fixes #33