Make FE model getAllCount() take language into account.
Since getAll() / getAllCount() are pretty much co-existing, it seems
logical that counting should also take the language column into account.

There is actually a much more efficient way to accomplish the limit,
offset retrieving + the total number of rows by using
`SQL_CALC_FOUND_ROWS` in the base query and then, right after that
executing `SELECT FOUND_ROWS()` but this requires changing the return
type of getAll() to a tuple, as well as construction the pagination data
structure in the client code.  For this, I did not have the time (yet
;-) ).